Fix FAS bug

+++ tests/userland/queue_mpmc_mcslock/mcslock.c
@@ -37,7 +37,7 @@ void mcs_unlock ( mcslock* L, qnode* I )
 	mcslock *succ;
 	if ( !(succ = I->next) )
 	{
-		if ( cmp_and_swap( L, (uint64_t)I, NULL) )
+		if ( cmp_and_swap( L, (uint64_t)I, (uint64_t)NULL) )
 			return;
 		do
 		{
